Feat/create edit roles and restructure phones

+++ src/config/axios.ts
@@ -1,4 +1,4 @@
-import axios from "axios"
+import axios, { AxiosError } from "axios"
 import { toast } from "react-toastify"
 
 const API_BASE_URL = 'https://api.fumiwo.io/v1/'
@@ -31,15 +31,12 @@ api.interceptors.response.use(
 		}
 		return response
 	},
-	(error) => {
-		if (error.response?.status === 401) {
+	(error: AxiosError) => {
+		const url = error.config?.url
+		if (error.response?.status === 401 && url?.includes("signin")) {
 			localStorage.clear()
 			toast.error("Session expired, please login again")
 			window.location.href = "/"
-		}
-
-		if (error.response?.status === 404) {
-			return Promise.resolve(error)
 		}
 
 		return Promise.reject(error)
